+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ## 2026-03-18
 - Implement some temporary fixes to FtrPath.  Not the best, but start relying on 
   DuckDb for settle prices.
+- Return null for the settlement price of an FTR auction that has not started to deliver.  
 
 ## 2026-03-17
 - Add pretty names to Caiso static buckets.